pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-codex-bridge"
version = "0.1.0"
description = "Installable Telegram bridge for local Codex sessions on macOS."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0,<22",
]

[project.optional-dependencies]
runtime = [
  "openai-whisper>=20240930",
]
dev = [
  "pytest>=7.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
